src/agents/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-pubsub
google-cloud-firestore
google-cloud-secret-manager
functions-framework
Flask
tenacity
crewai
langchain-community
langchain-core
langchain-anthropic
openai
tools